Refactor #293 - make TagId model
