+++ core/data-test/src/main/java/com/nlab/reminder/core/data/model/TagGenerator.kt
@@ -18,12 +18,14 @@ package com.nlab.reminder.core.data.model
 
 import com.nlab.testkit.faker.genBothify
 import com.nlab.testkit.faker.genInt
-import com.nlab.testkit.faker.genLong
+import com.nlab.testkit.faker.genLongGreaterThanZero
 
 /**
  * @author Doohyun
  */
+fun genTagId(value: Long = genLongGreaterThanZero()) = TagId(value)
 
-fun genTag(tagId: Long = genLong(), name: String = genBothify()) = Tag(tagId, name)
+fun genTag(id: TagId = genTagId(), name: String = genBothify()) = Tag(id, name)
+
 fun genTags(count: Int = genInt(min = 5, max = 10)): List<Tag> =
-    List(count) { index -> genTag(index.toLong()) }.distinctBy { it.name }
+    List(count) { index -> genTag(genTagId(index.toLong())) }.distinctBy { it.name }
